Returns no result for url string, missing "/"
Fixes bug where searching for a entry by its url slug would return no results. This happens because the url slug is always appended by a "/", but resolveRouteBinding searches without.

This fix checks if its a url slug or ID thats being used in the search. And if it is a slug, then appends wildcard characters